services/ingestion-capteurs/consumer/requirements.txt
aiokafka
pydantic
sqlalchemy>=2.0
asyncpg
python-dateutil
